Move getCostRangeIndex() to MVSpatialIndex
